+++ desktop/app/main.js
@@ -40,9 +40,6 @@ mb.on('after-create-window', () => {
 })
 
 mb.on('ready', () => {
-  require('../../react-native/react/native/pinentry').init()
-
-
   // prevent the menubar's window from dying when we quit
   mb.window.on('close', event => {
     mb.hideWindow()
@@ -75,7 +72,9 @@ installer(err => {
   splash()
 })
 
-app.dock.hide()
+if (app.dock) {
+  app.dock.hide()
+}
 
 // Don't quit the app, instead try to close all windows
 app.on('before-quit', event => {