+++ react-native/react/native/remote-manager.desktop.js
@@ -5,16 +5,23 @@ import {connect} from '../base-redux'
 
 import {bindActionCreators} from 'redux'
 import {registerIdentifyUi, onCloseFromHeader} from '../actions/tracker'
+import {registerPinentryListener, onCancel as pinentryOnCancel, onSubmit as pinentryOnSubmit} from '../actions/pinentry'
 // $FlowIssue platform files
 import RemoteComponent from './remote-component'
 
+import type {GUIEntryFeatures} from '../constants/types/flow-types'
 import type {TrackerState} from '../reducers/tracker'
+import type {PinentryState} from '../reducers/pinentry'
 
 export type RemoteManagerProps = {
+  registerPinentryListener: () => void,
+  pinentryOnCancel: (sessionID: number) => void,
+  pinentryOnSubmit: (sessionID: number, passphrase: string, features: GUIEntryFeatures) => void,
   registerIdentifyUi: () => void,
   onCloseFromHeader: () => void,
   trackerServerStarted: boolean,
-  trackers: {[key: string]: TrackerState}
+  trackers: {[key: string]: TrackerState},
+  pinentryStates: {[key: string]: PinentryState}
 }
 
 class RemoteManager extends Component {
@@ -31,6 +38,7 @@ class RemoteManager extends Component {
     if (!this.props.trackerServerStarted) {
       console.log('starting identify ui server')
       this.props.registerIdentifyUi()
+      this.props.registerPinentryListener()
     }
   }
 
@@ -43,6 +51,10 @@ class RemoteManager extends Component {
   shouldComponentUpdate (nextProps, nextState) {
     // different window states
     if (this.windowStates(nextProps.trackers) !== this.windowStates(this.props.trackers)) {
+      return true
+    }
+
+    if (nextProps.pinentryStates !== this.props.pinentryStates) {
       return true
     }
 
@@ -63,7 +75,7 @@ class RemoteManager extends Component {
               resizable: false
             }}
             waitForState
-            onRemoteClose={this.props.onCloseFromHeader}
+            onRemoteClose={() => this.props.onCloseFromHeader(username)}
             component='tracker'
             username={username}
             substore='tracker'
@@ -79,29 +91,64 @@ class RemoteManager extends Component {
     this.setState({popups})
   }
 
+  pinentryRemoteComponents () {
+    const windowOpts = {
+      width: 513, height: 250 + 20 /* TEMP workaround for header mouse clicks in osx */,
+      resizable: true,
+      fullscreen: false,
+      show: false,
+      frame: false
+    }
+
+    const {pinentryStates} = this.props
+
+    return Object.keys(pinentryStates).filter(sid => !pinentryStates[sid].closed).map((pSessionID) => {
+      const sid = parseInt(pSessionID, 10)
+      const onCancel = () => this.props.pinentryOnCancel(sid)
+      const onSubmit = this.props.pinentryOnSubmit.bind(null, sid)
+      return (
+        <RemoteComponent
+          windowsOpts={windowOpts}
+          waitForState
+          onRemoteClose={onCancel}
+          component='pinentry'
+          key={'pinentry:' + pSessionID}
+          onSubmit={onSubmit}
+          onCancel={onCancel}
+          sessionID={sid} />
+      )
+    })
+  }
+
   render () {
     return (
       <div>
-      {Object.keys(this.state.popups).filter(username => !this.props.trackers[username].closed).map(username => this.state.popups[username])}
+        {Object.keys(this.state.popups).filter(username => !this.props.trackers[username].closed).map(username => this.state.popups[username])}
+        {this.pinentryRemoteComponents()}
       </div>
     )
   }
 }
 
 RemoteManager.propTypes = {
+  pinentryOnCancel: React.PropTypes.any,
+  pinentryOnSubmit: React.PropTypes.any,
+  registerPinentryListener: React.PropTypes.any,
   registerIdentifyUi: React.PropTypes.any,
   onCloseFromHeader: React.PropTypes.any,
   trackerServerStarted: React.PropTypes.bool,
-  trackers: React.PropTypes.any
+  trackers: React.PropTypes.any,
+  pinentryStates: React.PropTypes.any
 }
 
 export default connect(
   state => {
     return {
       trackerServerStarted: state.tracker.serverStarted,
-      trackers: state.tracker.trackers
+      trackers: state.tracker.trackers,
+      pinentryStates: state.pinentry.pinentryStates || {}
     }
   },
-  dispatch => { return bindActionCreators({registerIdentifyUi, onCloseFromHeader}, dispatch) }
+  dispatch => bindActionCreators({registerIdentifyUi, onCloseFromHeader, registerPinentryListener, pinentryOnCancel, pinentryOnSubmit}, dispatch)
 )(RemoteManager)