+++ react-native/react/pinentry/index.render.desktop.js
@@ -6,7 +6,9 @@ import path from 'path'
 import commonStyles, {colors} from '../styles/common'
 
 export default class PinentryRender extends Component {
-  componentWillMount () {
+  constructor (props) {
+    super(props)
+
     this.state = {
       passphrase: '',
       features: {},
@@ -14,6 +16,10 @@ export default class PinentryRender extends Component {
     }
     for (const feature in this.props.features) {
       this.state.features[feature] = this.props.features[feature].defaultValue
+
+      if (feature === 'showTyping') {
+        this.state.showTyping = this.props.features[feature].defaultValue
+      }
     }
   }
 
@@ -36,7 +42,7 @@ export default class PinentryRender extends Component {
         <div style={styles.container}>
           <Header
             style={styles.header}
-            title='Passphrase needed'
+            title={this.props.windowTitle}
             onClose={() => this.props.onCancel()}
           />
           <div style={styles.bodyContainer}>
@@ -54,7 +60,7 @@ export default class PinentryRender extends Component {
                       name={feature}
                       value={feature}
                       label={this.props.features[feature].label}
-                      defaultChecked={this.props.features[feature].allow}
+                      defaultChecked={this.props.features[feature].defaultValue}
                       style={styles.checkbox}
                       onCheck={(_, checked) => this.onCheck(feature, checked)}/>
                   </div>
@@ -98,7 +104,6 @@ const styles = {
     backgroundColor: 'white',
     fontFamily: 'Noto Sans',
     fontSize: 15,
-    height: 184,
     width: 513
   },
   header: {
@@ -128,6 +133,7 @@ const styles = {
     alignItems: 'center',
     height: 49,
     paddingTop: 9,
+    paddingBottom: 9,
     paddingRight: 15
   },
   logo: {
@@ -143,7 +149,7 @@ const styles = {
     justifyContent: 'flex-end',
     position: 'absolute',
     right: 0,
-    top: 20
+    bottom: 55
   },
   checkbox: {
     marginTop: 30,