+++ go/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.6
+- libkb: Load optimizations; don't load unneeded fields (PR: keybase/client#1473)
+- engine: bugfix for user switching (PR: keybase/client#1474)
+- Fix goroutine leak in RPC calls (PR: keybase/client#1462)
+- Fix buggy "No device found No device found" error message
+- Fix PGP command line identify/track, flags (PR: keybase/client#1475)
+- Installer tweak: don't prompt to start service (windows)
+- Ansi color code support for terminal (Windows)
 ## 1.0.5-0 (2015-12-01)
 
 - Fix bug where cancelled RPC calls would cause hangs (PR: keybase/client#1433)